third_party/WebKit/LayoutTests/platform/win/fast/html/imports/import-add-child-to-change-order-of-dup-css-expected.txt

This is a testharness.js-based test.
PASS .target should initially be red 
PASS .target should be green after style-red.html is inserted as an import child before the inline green style. 
Harness: the test ran to completion.
